Revert "Merge pull request #108 from Akbi47/fix/media-compact-restore"
This reverts commit 2aa6a1150b9fd1a7320a3e7dc3b67513590fe1c9, reversing
changes made to 159dcdbc74c38af775728c8dfbb4e37e32f36aa8.
