--- src/features/blog/cache-tag.ts
@@ -1,1 +1,0 @@
-export const BLOG_CACHE_TAG = "blog";
