+++ src/features/cms/media-picker-modal.tsx
@@ -200,7 +200,6 @@ export function MediaPickerModal({
   }, [open]);
 
   if (!open) return null;
-  if (typeof document === "undefined") return null;
 
   function choose(item: MediaAsset) {
     onSelect({
