+++ src/styles/globals.css
@@ -5050,7 +5050,7 @@ video {
 
 .admin-media-grid {
   display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(8.5rem, 1fr));
+  grid-template-columns: repeat(auto-fill, minmax(10rem, 1fr));
   gap: 0.75rem;
   margin: 0;
   padding: 0;
@@ -5059,21 +5059,19 @@ video {
 
 .admin-media-grid__item {
   display: grid;
-  gap: 0.375rem;
-  padding: 0.5rem;
+  gap: 0.5rem;
+  padding: 0.625rem;
   border: 1px solid var(--color-border);
   border-radius: 0.75rem;
   background: var(--color-surface);
-  overflow: hidden;
 }
 
 .admin-media-grid__item .admin-media-preview {
   width: 100%;
-  aspect-ratio: 4 / 3;
+  height: 7rem;
   overflow: hidden;
   border-radius: 0.5rem;
   background: var(--color-surface-subtle);
-  cursor: pointer;
 }
 
 .admin-media-grid__item .admin-media-preview img {
@@ -5084,16 +5082,15 @@ video {
 
 .admin-media-grid__item .admin-media-meta {
   display: grid;
-  gap: 0.125rem;
+  gap: 0.25rem;
   min-width: 0;
-  padding: 0 0.125rem;
 }
 
 .admin-media-grid__item .admin-media-meta code {
   overflow: hidden;
   text-overflow: ellipsis;
   white-space: nowrap;
-  font-size: 0.6875rem;
+  font-size: 0.75rem;
 }
 
 .admin-media-title {
@@ -5106,13 +5103,6 @@ video {
 
 .admin-media-dims {
   color: var(--color-text-muted);
-  font-size: 0.6875rem;
-}
-
-.admin-media-actions {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.25rem 0.625rem;
   font-size: 0.75rem;
 }
 
@@ -5127,38 +5117,6 @@ video {
 .admin-pagination__info {
   color: var(--color-text-muted);
   font-size: 0.875rem;
-}
-
-/* Image lightbox inside the media library (view a large version on click). */
-.admin-lightbox {
-  width: min(56rem, calc(100vw - 2rem));
-}
-
-.admin-lightbox__image {
-  display: block;
-  width: 100%;
-  max-height: 62vh;
-  object-fit: contain;
-  border-radius: 0.5rem;
-  background: var(--color-surface-subtle);
-  margin-block-end: 0.75rem;
-}
-
-.admin-lightbox__meta {
-  display: grid;
-  gap: 0.25rem;
-  margin-block-end: 0.75rem;
-}
-
-.admin-lightbox__meta code {
-  overflow-wrap: anywhere;
-  color: var(--color-text-muted);
-  font-size: 0.75rem;
-}
-
-.admin-lightbox__meta span {
-  color: var(--color-text-muted);
-  font-size: 0.8125rem;
 }
 
 /* --- Shared MediaPickerModal --- */
